fix: The derives stop guessing: strict attributes, honest types, Seaography gone
Every #[pgorm(...)] parser now rejects unknown keys with a spanned
error naming the offender — the typo that silently turned bigint into
bigserial is a compile error, matching the posture DeriveValueType
already had. skip accumulates instead of being recomputed away by the
next meta item; DerivePartialModel accumulates instead of clobbering,
making the both-keys conflict error reachable for the first time; and
the type-inference table drops the i8, u32, u64 and char rows that
advertised column types the decode layer could never read back.

FromJsonQueryResult serialization failure now panics naming the type
instead of silently binding NULL over stored data. The crate-rename
alias workaround is documented where users hit it.

Seaography is deleted entirely per the recorded operator decision:
the feature, DeriveRelatedEntity, the codegen option and its golden
directory, the fixture — a feature that could never compile is no
longer advertised, and its spec rule is retired with the rationale
kept as negative space. The inherited sea_attr locals and the
sea_orm_active_enums fixture lose their ancestral names with it.

nplan-id: macro-unknown-attrs
nplan-id: macro-from-query-result-skip
nplan-id: macro-partial-model-attrs
nplan-id: macro-model-type-inference
nplan-id: macro-dependency-renames
nplan-id: seaography-feature-build
nplan-id: from-json-serialization
nplan-id: macro-sea-attr-name
nplan-id: sea-orm-active-enums-name
nplan-rule: [spec:pgorm:def:codegen.entity]
nplan-rule: [spec:pgorm:def:macros.derive]
nplan-rule: [spec:pgorm:sem:codegen.entity.expanded.blocks]
nplan-rule: [spec:pgorm:sem:codegen.entity.seaography]
nplan-rule: [spec:pgorm:sem:macros.derive.entity-model.column-def]
nplan-rule: [spec:pgorm:sem:macros.derive.from-query-result]
nplan-rule: [spec:pgorm:sem:macros.derive.partial-model]
nplan-rule: [spec:pgorm:sem:macros.derive.value-type]
nplan-rule: [spec:pgorm:syn:macros.derive.entity-model.attrs]
nplan-rule: [spec:pgorm:sem:macros.derive.entity-model.column-def/test]
nplan-rule: [spec:pgorm:sem:macros.derive.partial-model/test]

+++ src/entity/column.rs
@@ -595,7 +595,7 @@ mod tests {
                 pub id: i32,
                 pub one: i32,
                 #[pgorm(unique)]
-                pub two: i8,
+                pub two: i16,
                 #[pgorm(indexed)]
                 pub three: i16,
                 #[pgorm(nullable)]
@@ -603,11 +603,11 @@ mod tests {
                 #[pgorm(unique, indexed, nullable)]
                 pub five: i64,
                 #[pgorm(nullable)]
-                pub eight: u32,
+                pub eight: i64,
                 #[pgorm(default_expr = "Expr::current_timestamp()")]
                 pub ten: DateTimeUtc,
                 #[pgorm(default_value = 7)]
-                pub eleven: i8,
+                pub eleven: i16,
                 #[pgorm(default_value = "twelve_value")]
                 pub twelve: String,
                 #[pgorm(default_expr = "\"twelve_value\"")]
